feat(github-comments): ✨ addressed github comments
